+++ src/mock/constant.js
@@ -1,2 +1,96 @@
 export const POINT_COUNT = 3;
 export const EDIT_COUNT = 1;
+export const images = [1, 2, 3, 4, 5];
+
+export const offersItems = [
+  {
+    checked: `checked=""`,
+    id: `luggage-1`,
+    name: `Add luggage`,
+    price: 30
+  },
+  {
+    checked: ``,
+    id: `comfort-1`,
+    name: `Switch to comfort class`,
+    price: 100
+  },
+  {
+    checked: ``,
+    id: `meal-1`,
+    mame: `Add meal`,
+    price: 15
+  },
+  {
+    checked: ``,
+    id: `seats-1`,
+    name: `Choose seats`,
+    price: 5
+  },
+  {
+    checked: ``,
+    id: `train-1`,
+    name: `Travel by train`,
+    price: 40
+  }
+];
+
+export const transports = [
+  {
+    id: `taxi`,
+    name: `Taxi`
+  },
+  {
+    id: `bus`,
+    name: `Bus`
+  },
+  {
+    id: `train`,
+    name: `train`
+  },
+  {
+    id: `ship`,
+    name: `Ship`
+  },
+  {
+    id: `transport`,
+    name: `Transport`
+  },
+  {
+    id: `flight`,
+    name: `Flight`
+  },
+  {
+    id: `check-in`,
+    name: `Check-in`
+  },
+  {
+    id: `sightseeing`,
+    name: `Sightseeing`
+  },
+  {
+    id: `restaurant`,
+    name: `Restaurant`
+  }];
+export const cities = [`Amsterdam`, `Geneva`, `Chamonix`];
+export const filters = [
+  {
+    checked: `checked=""`,
+    name: `Day`
+  },
+  {
+    checked: ``,
+    name: `Event`
+  },
+  {
+    checked: ``,
+    name: `Time`
+  },
+  {
+    checked: ``,
+    name: `Price`
+  },
+  {
+    checked: ``,
+    name: `Offers`
+  }];
